requirements.txt
discord.py>=2.3.0
yt-dlp>=2024.1.0
spotipy>=2.23.0
lyricsgenius>=3.0.1
PyNaCl>=1.5.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
